Avoid circular imports by moving ForwardMode to decomon.core
