+++ src/decomon/layers/__init__.py
@@ -1,3 +1,2 @@
-from ..core import InputsOutputsSpec
-from .core import ForwardMode
+from ..core import ForwardMode, InputsOutputsSpec
 from .utils import get_lower, get_upper, maximum, minimum
