+++ src/decomon/layers/core.py
@@ -1,39 +1,13 @@
 from abc import ABC, abstractmethod
-from enum import Enum
 from typing import Any, Dict, List, Optional, Type, Union
 
 import tensorflow as tf
 from tensorflow.keras.layers import Layer
 
-from decomon.core import BoxDomain, InputsOutputsSpec, PerturbationDomain
+from decomon.core import BoxDomain, ForwardMode, InputsOutputsSpec, PerturbationDomain
 from decomon.keras_utils import get_weight_index_from_name
 
-
-class ForwardMode(Enum):
-    """The different forward (from input to output) linear based relaxation perturbation analysis."""
-
-    IBP = "ibp"
-    """Propagation of constant bounds from input to output."""
-
-    AFFINE = "affine"
-    """Propagation of affine bounds from input to output."""
-
-    HYBRID = "hybrid"
-    """Propagation of constant and affines bounds from input to output."""
-
-
 DEEL_LIP = "deel-lip>"
-
-
-def get_mode(ibp: bool = True, affine: bool = True) -> ForwardMode:
-
-    if ibp:
-        if affine:
-            return ForwardMode.HYBRID
-        else:
-            return ForwardMode.IBP
-    else:
-        return ForwardMode.AFFINE
 
 
 class DecomonLayer(ABC, Layer):
