+++ src/decomon/models/convert.py
@@ -9,9 +9,8 @@ from tensorflow.keras.models import Model
 
 from decomon.backward_layers.core import BackwardLayer
 from decomon.backward_layers.crown import Convert2Mode
-from decomon.core import BoxDomain, PerturbationDomain, Slope
+from decomon.core import BoxDomain, PerturbationDomain, Slope, get_mode
 from decomon.layers.convert import to_decomon
-from decomon.layers.core import get_mode
 from decomon.models.backward_cloning import convert_backward
 from decomon.models.forward_cloning import (
     LayerMapDict,
