+++ tests/test_activation.py
@@ -6,9 +6,8 @@ import pytest
 import tensorflow.keras.backend as K
 from numpy.testing import assert_almost_equal
 
-from decomon.core import Slope
+from decomon.core import ForwardMode, Slope
 from decomon.layers.activations import relu, sigmoid, softmax, softsign, tanh
-from decomon.layers.core import ForwardMode
 
 
 @pytest.mark.parametrize(
