+++ tests/test_clone.py
@@ -8,8 +8,7 @@ from tensorflow.keras.layers import Activation, Dense, Flatten, Input, Reshape
 from tensorflow.keras.models import Sequential
 
 from decomon.backward_layers.utils import get_affine, get_ibp
-from decomon.core import Slope
-from decomon.layers.core import ForwardMode, get_mode
+from decomon.core import ForwardMode, Slope, get_mode
 from decomon.layers.decomon_reshape import DecomonReshape
 from decomon.models import clone
 from decomon.models.convert import FeedDirection, get_direction
