+++ tests/test_clone_forward.py
@@ -16,7 +16,7 @@ from tensorflow.keras.layers import (
 from tensorflow.keras.models import Model, Sequential
 
 from decomon.backward_layers.utils import get_affine, get_ibp
-from decomon.layers.core import ForwardMode
+from decomon.core import ForwardMode
 from decomon.models.forward_cloning import convert_forward
